refactor out CommentBox, add it to userside
